合并1.0.5版本的代码

+++ fox-edge-server-protocol/fox-edge-server-protocol-dlt645-core/src/main/java/cn/foxtech/device/protocol/v1/dlt645/core/DLT645Protocol.java
@@ -272,7 +272,7 @@ public class DLT645Protocol {
             return null;
         }
 
-        byte byFun = Byte.decode(param.get(FUN).toString());
+        byte byFun = Integer.decode(param.get(FUN).toString()).byteValue();
         byte[] byAddr = (byte[]) (param.get(ADR));
         byte[] arrData = (byte[]) (param.get(DAT));
